+++ server_administration.go
@@ -299,15 +299,36 @@ type changelogTemplateVars struct {
 }
 
 func (sah *ServerAdministrationHandler) changelog(w http.ResponseWriter, r *http.Request) {
-	changelog, err := LoadChangelog()
+	sah.viewRenderer.MustLoadTemplate(w, r, "changelog.html", &changelogTemplateVars{
+		Changelog: Changelog,
+	})
+}
+
+func (sah *ServerAdministrationHandler) robots(w http.ResponseWriter, r *http.Request) {
+	// do we want to let robots on the internet know things about us?!?
+	serverOpts, err := sah.store.LoadServerOptions()
 
 	if err != nil {
-		logrus.WithError(err).Error("could not load changelog")
+		logrus.WithError(err).Errorf("couldn't load server options")
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 		return
 	}
 
-	sah.viewRenderer.MustLoadTemplate(w, r, "changelog.html", &changelogTemplateVars{
-		Changelog: changelog,
-	})
+	var response string
+
+	w.Header().Set("Content-Type", "text/plain")
+
+	if serverOpts.PreventWebCrawlers == 1 {
+		response = "User-agent: *\nDisallow: /"
+	} else {
+		response = "User-agent: *\nDisallow:"
+	}
+
+	_, err = w.Write([]byte(response))
+
+	if err != nil {
+		logrus.WithError(err).Errorf("couldn't write response text")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
 }