+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Added:
 Fixes:
 
 * Fixes track/car display in dark mode
+* Added an option to prevent web crawlers from accessing any pages on the manager using a robots.txt.
+* Fixes track details page names to be a bit nicer
+* Added Black Cat County to the list of default content
+* Fixes an issue where Server Manager would not start when a recurring race with an end date had a scheduled recurrence while Server Manager was offline.
+
 
 ---